requirements.txt
faiss-cpu
transformers
torch
gradio
accelerate
bitsandbytes
